Restructure Duke project and incooperate gradle properly
