+++ src/main/java/duke/task/TaskList.java
@@ -1,4 +1,4 @@
-package java.task;
+package duke.task;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
