+++ text-ui-test/input.txt
@@ -17,7 +17,7 @@ deadline return book /by 2016-12-25 18:00
 deadline return book /by 2016-12-25
 deadline return book
 event project meeting /at 2016-12-25
-event project meeting /at 2016-12-25 18:00
-event project meeting /at 2016-12-25 18:00 19:00
+eventevent project meeting /at 2016-12-25 18:00 19:00 project meeting /at 2016-12-25 18:00
+
 event project meeting /at 2016-12-25 19:00 18:00
 event project meeting /at 2016-12-25 19:00 19:00
